@@ -1,20 +1,18 @@
-cmake_minimum_required(VERSION 3.0.2)
+cmake_minimum_required(VERSION 3.5S)
 project(apriltag_ros)
 
-find_package(catkin REQUIRED COMPONENTS
-  cmake_modules
-  cv_bridge
-  geometry_msgs
-  image_geometry
-  image_transport
-  message_generation
-  nodelet
-  pluginlib
-  roscpp
-  sensor_msgs
-  std_msgs
-  tf
-)
+### Compile options
+if(NOT CMAKE_CXX_STANDARD)
+  set(CMAKE_CXX_STANDARD 14)
+  set(CMAKE_CXX_STANDARD_REQUIRED ON)
+  set(CMAKE_CXX_EXTENSIONS OFF)
+endif()
+if(CMAKE_COMPILER_IS_GNUCXX OR CMAKE_CXX_COMPILER_ID MATCHES "Clang")
+  add_compile_options(-Wall -Wextra -Wpedantic -Werror)
+endif()
+
+find_package(ament_cmake_auto REQUIRED)
+ament_auto_find_build_dependencies()
 
 find_package(Eigen3 REQUIRED)
 find_package(OpenCV REQUIRED)
