@@ -1,34 +1,38 @@
+// Copyright (c) 2017, California Institute of Technology.
+// All rights reserved.
+//
+// Software License Agreement (BSD 2-Clause Simplified License)
+//
+// Redistribution and use in source and binary forms, with or without
+// modification, are permitted provided that the following conditions
+// are met:
+//
+//  * Redistributions of source code must retain the above copyright
+//    notice, this list of conditions and the following disclaimer.
+//  * Redistributions in binary form must reproduce the above
+//    copyright notice, this list of conditions and the following
+//    disclaimer in the documentation and/or other materials provided
+//    with the distribution.
+//
+// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+// "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+// LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
+// FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE
+// COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
+// INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
+// BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+// LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+// CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
+// LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
+// ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+// POSSIBILITY OF SUCH DAMAGE.
+//
+// The views and conclusions contained in the software and documentation are
+// those of the authors and should not be interpreted as representing official
+// policies, either expressed or implied, of the California Institute of
+// Technology.
 /**
- * Copyright (c) 2017, California Institute of Technology.
- * All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- *
- * 1. Redistributions of source code must retain the above copyright notice,
- *    this list of conditions and the following disclaimer.
- * 2. Redistributions in binary form must reproduce the above copyright notice,
- *    this list of conditions and the following disclaimer in the documentation
- *    and/or other materials provided with the distribution.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
- * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
- * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
- * ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE
- * LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
- * CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
- * SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
- * INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
- * CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
- * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
- * POSSIBILITY OF SUCH DAMAGE.
- *
- * The views and conclusions contained in the software and documentation are
- * those of the authors and should not be interpreted as representing official
- * policies, either expressed or implied, of the California Institute of
- * Technology.
- *
- ** common_functions.h *********************************************************
+ ** common_functions.hpp *******************************************************
  *
  * Wrapper classes for AprilTag standalone and bundle detection. Main function
  * is TagDetector::detectTags which wraps the call to core AprilTag 2
@@ -41,67 +45,61 @@
  * Originator:        Danylo Malyuta, JPL
  ******************************************************************************/
 
-#ifndef APRILTAG_ROS_COMMON_FUNCTIONS_H
-#define APRILTAG_ROS_COMMON_FUNCTIONS_H
+#ifndef APRILTAG_ROS__COMMON_FUNCTIONS_HPP_
+#define APRILTAG_ROS__COMMON_FUNCTIONS_HPP_
+
+#include <apriltag.h>
 
 #include <string>
 #include <sstream>
 #include <vector>
 #include <map>
+#include <memory>
 
-#include <ros/ros.h>
-#include <ros/console.h>
-#include <XmlRpcException.h>
-#include <cv_bridge/cv_bridge.h>
-#include <eigen3/Eigen/Dense>
-#include <eigen3/Eigen/Geometry>
-#include <opencv2/opencv.hpp>
-#include <opencv2/imgproc/imgproc.hpp>
-#include <opencv2/core/core.hpp>
-#include <image_transport/image_transport.h>
-#include <sensor_msgs/image_encodings.h>
-#include <tf/transform_broadcaster.h>
+#include "eigen3/Eigen/Dense"
+#include "eigen3/Eigen/Geometry"
 
-#include <apriltag.h>
+#include "opencv2/opencv.hpp"
+#include "opencv2/imgproc/imgproc.hpp"
+#include "opencv2/core/core.hpp"
 
-#include "apriltag_ros/AprilTagDetection.h"
-#include "apriltag_ros/AprilTagDetectionArray.h"
+#include "rclcpp/rclcpp.hpp"
+#include "cv_bridge/cv_bridge.h"
+#include "geometry_msgs/msg/pose_with_covariance_stamped.hpp"
+#include "image_transport/image_transport.hpp"
+#include "sensor_msgs/image_encodings.hpp"
+#include "tf2_ros/transform_broadcaster.h"
+
+#include "apriltag_msgs/msg/april_tag_detection.hpp"
+#include "apriltag_msgs/msg/april_tag_detection_array.hpp"
 
 namespace apriltag_ros
 {
 
-template<typename T>
-T getAprilTagOption(ros::NodeHandle& pnh,
-                    const std::string& param_name, const T & default_val)
-{
-  T param_val;
-  pnh.param<T>(param_name, param_val, default_val);
-  return param_val;
-}
-
 // Stores the properties of a tag member of a bundle
 struct TagBundleMember
 {
-  int id; // Payload ID
-  double size; // [m] Side length
-  cv::Matx44d T_oi; // Rigid transform from tag i frame to bundle origin frame
+  int id;  // Payload ID
+  double size;  // [m] Side length
+  cv::Matx44d T_oi;  // Rigid transform from tag i frame to bundle origin frame
 };
 
 class StandaloneTagDescription
 {
- public:
-  StandaloneTagDescription() {};
-  StandaloneTagDescription(int id, double size,
-                           std::string &frame_name) :
-      id_(id),
-      size_(size),
-      frame_name_(frame_name) {}
+public:
+  StandaloneTagDescription() {}
+  StandaloneTagDescription(
+    int id, double size,
+    const std::string & frame_name)
+  : id_(id),
+    size_(size),
+    frame_name_(frame_name) {}
 
-  double size() { return size_; }
-  int id() { return id_; }
-  std::string& frame_name() { return frame_name_; }
+  double size() {return size_;}
+  int id() {return id_;}
+  std::string & frame_name() {return frame_name_;}
 
- private:
+private:
   // Tag description
   int id_;
   double size_;
@@ -110,24 +108,26 @@ class StandaloneTagDescription
 
 class TagBundleDescription
 {
- public:
-  std::map<int, int > id2idx_; // (id2idx_[<tag ID>]=<index in tags_>) mapping
+public:
+  std::map<int, int> id2idx_;  // (id2idx_[<tag ID>]=<index in tags_>) mapping
 
-  TagBundleDescription(std::string name) :
-      name_(name) {}
+  explicit TagBundleDescription(std::string name)
+  : name_(name) {}
 
-  void addMemberTag(int id, double size, cv::Matx44d T_oi) {
+  void addMemberTag(int id, double size, cv::Matx44d T_oi)
+  {
     TagBundleMember member;
     member.id = id;
     member.size = size;
     member.T_oi = T_oi;
     tags_.push_back(member);
-    id2idx_[id] = tags_.size()-1;
+    id2idx_[id] = tags_.size() - 1;
   }
 
-  std::string name () const { return name_; }
+  std::string name() const {return name_;}
   // Get IDs of bundle member tags
-  std::vector<int> bundleIds () {
+  std::vector<int> bundleIds()
+  {
     std::vector<int> ids;
     for (unsigned int i = 0; i < tags_.size(); i++) {
       ids.push_back(tags_[i].id);
@@ -135,28 +135,29 @@ class TagBundleDescription
     return ids;
   }
   // Get sizes of bundle member tags
-  std::vector<double> bundleSizes () {
+  std::vector<double> bundleSizes()
+  {
     std::vector<double> sizes;
     for (unsigned int i = 0; i < tags_.size(); i++) {
       sizes.push_back(tags_[i].size);
     }
     return sizes;
   }
-  int memberID (int tagID) { return tags_[id2idx_[tagID]].id; }
-  double memberSize (int tagID) { return tags_[id2idx_[tagID]].size; }
-  cv::Matx44d memberT_oi (int tagID) { return tags_[id2idx_[tagID]].T_oi; }
+  int memberID(int tagID) {return tags_[id2idx_[tagID]].id;}
+  double memberSize(int tagID) {return tags_[id2idx_[tagID]].size;}
+  cv::Matx44d memberT_oi(int tagID) {return tags_[id2idx_[tagID]].T_oi;}
 
- private:
+private:
   // Bundle description
   std::string name_;
-  std::vector<TagBundleMember > tags_;
+  std::vector<TagBundleMember> tags_;
 };
 
 class TagDetector
 {
- private:
+private:
   // Detections sorting
-  static int idComparison(const void* first, const void* second);
+  static int idComparison(const void * first, const void * second);
 
   // Remove detections of tags with the same ID
   void removeDuplicates();
@@ -173,47 +174,43 @@ class TagDetector
                                   // you want the largest value possible.
 
   // AprilTag 2 objects
-  apriltag_family_t *tf_;
-  apriltag_detector_t *td_;
-  zarray_t *detections_;
+  apriltag_family_t * tf_;
+  apriltag_detector_t * td_;
+  zarray_t * detections_;
+
+  // Node objects
+  rclcpp::Clock::SharedPtr clock_;
+  rclcpp::Logger logger_{rclcpp::get_logger("tag_detector")};
 
   // Other members
   std::map<int, StandaloneTagDescription> standalone_tag_descriptions_;
-  std::vector<TagBundleDescription > tag_bundle_descriptions_;
+  std::vector<TagBundleDescription> tag_bundle_descriptions_;
   bool remove_duplicates_;
   bool run_quietly_;
   bool publish_tf_;
-  tf::TransformBroadcaster tf_pub_;
+  std::unique_ptr<tf2_ros::TransformBroadcaster> tf_pub_;
 
- public:
-
-  TagDetector(ros::NodeHandle pnh);
+public:
+  explicit TagDetector(rclcpp::Node * node);
   ~TagDetector();
 
   // Store standalone and bundle tag descriptions
-  std::map<int, StandaloneTagDescription> parseStandaloneTags(
-      XmlRpc::XmlRpcValue& standalone_tag_descriptions);
-  std::vector<TagBundleDescription > parseTagBundles(
-      XmlRpc::XmlRpcValue& tag_bundles);
-  double xmlRpcGetDouble(
-      XmlRpc::XmlRpcValue& xmlValue, std::string field) const;
-  double xmlRpcGetDoubleWithDefault(
-      XmlRpc::XmlRpcValue& xmlValue, std::string field,
-      double defaultValue) const;
+  std::map<int, StandaloneTagDescription> parseStandaloneTags(rclcpp::Node * node);
+  std::vector<TagBundleDescription> parseTagBundles(rclcpp::Node * node);
 
   bool findStandaloneTagDescription(
-      int id, StandaloneTagDescription*& descriptionContainer,
-      bool printWarning = true);
+    int id, StandaloneTagDescription * & descriptionContainer,
+    bool printWarning = true);
 
-  geometry_msgs::PoseWithCovarianceStamped makeTagPose(
-      const Eigen::Matrix4d& transform,
-      const Eigen::Quaternion<double> rot_quaternion,
-      const std_msgs::Header& header);
+  geometry_msgs::msg::PoseWithCovarianceStamped makeTagPose(
+    const Eigen::Matrix4d & transform,
+    const Eigen::Quaternion<double> rot_quaternion,
+    const std_msgs::msg::Header & header);
 
   // Detect tags in an image
-  AprilTagDetectionArray detectTags(
-      const cv_bridge::CvImagePtr& image,
-      const sensor_msgs::CameraInfoConstPtr& camera_info);
+  apriltag_msgs::msg::AprilTagDetectionArray detectTags(
+    const cv_bridge::CvImagePtr & image,
+    const sensor_msgs::msg::CameraInfo::ConstSharedPtr & camera_info);
 
   // Get the pose of the tag in the camera frame
   // Returns homogeneous transformation matrix [R,t;[0 0 0 1]] which
@@ -223,21 +220,23 @@ class TagDetector
   // vector from the camera frame origin to the tag frame origin,
   // expressed in the camera frame.
   Eigen::Matrix4d getRelativeTransform(
-      std::vector<cv::Point3d > objectPoints,
-      std::vector<cv::Point2d > imagePoints,
-      double fx, double fy, double cx, double cy) const;
-  
-  void addImagePoints(apriltag_detection_t *detection,
-                      std::vector<cv::Point2d >& imagePoints) const;
-  void addObjectPoints(double s, cv::Matx44d T_oi,
-                       std::vector<cv::Point3d >& objectPoints) const;
+    std::vector<cv::Point3d> objectPoints,
+    std::vector<cv::Point2d> imagePoints,
+    double fx, double fy, double cx, double cy) const;
+
+  void addImagePoints(
+    apriltag_detection_t * detection,
+    std::vector<cv::Point2d> & imagePoints) const;
+  void addObjectPoints(
+    double s, cv::Matx44d T_oi,
+    std::vector<cv::Point3d> & objectPoints) const;
 
   // Draw the detected tags' outlines and payload values on the image
   void drawDetections(cv_bridge::CvImagePtr image);
 
-  bool get_publish_tf() const { return publish_tf_; }
+  bool get_publish_tf() const {return publish_tf_;}
 };
 
-} // namespace apriltag_ros
+}  // namespace apriltag_ros
 
-#endif // APRILTAG_ROS_COMMON_FUNCTIONS_H
+#endif  // APRILTAG_ROS__COMMON_FUNCTIONS_HPP_
